requirements.txt
Django==1.11
falcon==1.1.0
gunicorn==19.7.0
ipdb==0.10.2
ipython==5.3.0
marshmallow==2.13.1
ujson==1.35
django-extensions==1.7.7
django-model-utils==2.6.1
Pillow==4.0.0
psycopg2==2.7
